# docs/swagger.yaml
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  pager.PageModel:
    properties:
      count:
        type: integer
      list: {}
    type: object
...
  schema.GetVoteWithPageResp:
    properties:
      answer_id:
        description: answer id
        type: string
      content:
        description: content
        type: string
      created_at:
        description: create time
        type: integer
      object_id:
        description: object id
        type: string
      object_type:
        description: object type
        enum:
        - question
        - answer
        - tag
        - comment
        type: string
      question_id:
        description: question id
        type: string
      title:
        description: title
        type: string
      url_title:
        description: url title
        type: string
      vote_type:
        description: vote type
        type: string
    type: object
...
  schema.VoteReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      is_cancel:
        type: boolean
      object_id:
        type: string
    required:
    - object_id
    type: object
  schema.VoteResp:
    properties:
      down_votes:
        type: integer
      up_votes:
        type: integer
      vote_status:
        type: string
      votes:
        type: integer
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/personal/vote/page:
    get:
      consumes:
      - application/json
      description: get user personal votes
      parameters:
      - description: page size
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.GetVoteWithPageResp'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: get user personal votes
      tags:
      - Activity
...
  /answer/api/v1/vote/down:
    post:
      consumes:
      - application/json
      description: add vote
      parameters:
      - description: vote
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.VoteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.VoteResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: vote down
      tags:
      - Activity
  /answer/api/v1/vote/up:
    post:
      consumes:
      - application/json
      description: add vote
      parameters:
      - description: vote
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.VoteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.VoteResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: vote up
      tags:
      - Activity
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"